requirements.txt
selenium==4.3.0
pyotp==2.6.0
cacheout==0.13.1
python-dotenv==0.20.0
boto3==1.24.28
# retrying==1.3.3
retry==0.9.2